Move Script into module package
